+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [1.1.0] / 2025-05-16
+### Features
+- Support custom `Element` and `Attribute`.
+### PackageBuilder
+- Add `ModelBase` with custom `IXmlSerializable`.
+- Add `ModelBaseExtension` with `CreateEntryElement` and `CreateAttribute`.
+
 ## [1.0.6] / 2023-11-24 - 2023-12-06
 ### Features
 - `Tests` project.
@@ -56,6 +63,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - First Release
 
 [vNext]: ../../compare/1.0.0...HEAD
+[1.1.0]: ../../compare/1.0.6...1.1.0
 [1.0.6]: ../../compare/1.0.5...1.0.6
 [1.0.5]: ../../compare/1.0.4...1.0.5
 [1.0.4]: ../../compare/1.0.3...1.0.4